+++ e2e_tests/fixtures/build-default/src/client/index.js
@@ -1,2 +1,6 @@
 
+import '../public/img.jpg';
+import '../public/script.js';
+import '../assets/file.ico';
+
 module.exports = { client: true };